+++ src/dashpi.py
@@ -31,9 +31,8 @@ from blueprints.apikeys import apikeys_bp
 from blueprints.loops import loops_bp
 from blueprints.wifi import wifi_bp
 from blueprints.auth import auth_bp
-from plugins.pluginmanager.api import plugin_manage_bp
 from jinja2 import ChoiceLoader, FileSystemLoader
-from plugins.plugin_registry import load_plugins
+from plugins.plugin_registry import load_plugins, register_plugin_blueprints
 from waitress import serve
 
 
@@ -147,7 +146,7 @@ app.register_blueprint(apikeys_bp)
 app.register_blueprint(loops_bp)
 app.register_blueprint(wifi_bp)
 app.register_blueprint(auth_bp)
-app.register_blueprint(plugin_manage_bp)
+register_plugin_blueprints(app)
 
 # Inject project_name and version into all templates
 @app.context_processor